able to update customers

+++ app.js
@@ -22,7 +22,7 @@ const crmMenu = async () => {
       await viewAllCustomers();
       break;
     case "3":
-      await updateCustomers();
+      await updateCustomer();
       break;
     case "4":
       await deleteCustomer();
@@ -49,12 +49,12 @@ connect();
 
 const viewAllCustomers = async () => {
   const allCustomers = await Customers.find();
-  if (!allCustomers.length === 0) {
+  if (allCustomers.length !== 0) {
     console.log("All customer:", allCustomers);
     crmMenu();
   } else {
     console.log("There are no customers. Please add some to the CRM");
-    crmMenu();
+    await crmMenu();
   }
 };
 
@@ -65,15 +65,34 @@ const createCustomer = async () => {
   };
   const customer = await Customers.create(customerData);
   console.log("New customer:", customer);
-  crmMenu();
+  await crmMenu();
 };
 
-const updateCustomers = async () => {};
+const updateCustomer = async () => {
+  console.log(`Below is a list of customers:\n`, await Customers.find());
+
+  const customerId = prompt(
+    `Copy and paste the id of the customer you would like to update here:\n`
+  );
+  console.log(`# user inputs ${customerId}`);
+  const customer = await Customers.findByIdAndUpdate(customerId, {
+    Name: prompt(`What is the customer's new name?\n`),
+    Age: prompt(`What is the customer's new age?\n`),
+  });
+  console.log("Details updated:", customer);
+  await crmMenu();
+};
 
 const deleteCustomer = async () => {
   const customerId = prompt("Enter the customer's id: ");
-
-  const customer = await Customers.findByIdAndDelete(customerId);
-  console.log("Deleted customer:", customer);
-  crmMenu();
+  const customer = await Customers.findById(customerId);
+  if (customer) {
+    await Customers.findByIdAndDelete(customerId);
+    console.log("Deleted customer:", customer);
+    await crmMenu();
+    ``;
+  } else {
+    console.log("Customer not found");
+    await crmMenu();
+  }
 };
